fix: keep trade positions out of startup graph (OK-59276)
