+++ packages/kit/src/states/jotai/contexts/swap/actions.ts
@@ -13,14 +13,16 @@ import {
   shouldShowSwapAccountUnsupportedAlert,
 } from '@onekeyhq/kit/src/views/Swap/utils/swapNoWalletWarningGuard';
 import {
-  SWAP_PRO_POSITIONS_RUNTIME_TTL_MS,
   getValidSwapProPositionsCache,
   upsertSwapProPositionsCacheEntry,
   upsertSwapProPositionsRuntimeEntry,
 } from '@onekeyhq/kit/src/views/Swap/utils/swapProPositionsCacheUtils';
+import type {
+  ISwapProPositionsLoadOptions,
+  ISwapProPositionsLoader,
+} from '@onekeyhq/kit/src/views/Swap/utils/swapProPositionsLoader';
 import type { ISwapProTokenCarryUtils } from '@onekeyhq/kit/src/views/Swap/utils/swapProTokenCarryUtils';
 import { buildSwapRateDifference } from '@onekeyhq/kit/src/views/Swap/utils/swapRateDifferenceUtils';
-import { buildStockPositionTokens } from '@onekeyhq/kit/src/views/Swap/utils/swapStockPositionsUtils';
 import { moveNetworkToFirst } from '@onekeyhq/kit/src/views/Swap/utils/utils';
 import {
   currencyPersistAtom,
@@ -54,7 +56,6 @@ import {
 import {
   getSwapBridgeDefaultToToken,
   swapDefaultSetTokens,
-  swapProStockPositionsListMinValue,
   swapQuoteIntervalMaxCount,
   swapRefreshInterval,
   swapStockTokenListMaxCount,
@@ -97,7 +98,6 @@ import {
   type ISwapInputAmountDraft,
   type ISwapQuoteEventErrorState,
   type ISwapTokenAmountState,
-  buildSwapProPositionsOwnerKey,
   contextAtomMethod,
   limitOrderMarketPriceAtom,
   rateDifferenceAtom,
@@ -120,8 +120,6 @@ import {
   swapProDirectionAtom,
   swapProInputAmountAtom,
   swapProPositionsCacheAtom,
-  swapProPositionsRequestIdAtom,
-  swapProPositionsRequestIdsAtom,
   swapProPositionsRuntimeDataAtom,
   swapProSelectTokenAtom,
   swapProSellToTokenAtom,
@@ -3116,382 +3114,27 @@ class ContentJotaiActionsSwap extends ContextJotaiActionsBase {
   );
 
   swapProLoadSupportNetworksTokenList = contextAtomMethod(
-    async (
+    (
       get,
       set,
       supportNetworks: ISwapNetwork[],
       indexedAccountId?: string,
       otherWalletTypeAccountId?: string,
       currencyId?: string,
-      options?: {
-        forceRefresh?: boolean;
-        stockOnly?: boolean;
-        additionalSupportNetworkScopes?: {
-          stockOnly?: boolean;
-          supportNetworks: ISwapNetwork[];
-        }[];
+      options?: ISwapProPositionsLoadOptions & {
+        positionLoader?: ISwapProPositionsLoader;
       },
-    ) => {
-      const positionCurrencyId = currencyId?.toLowerCase() ?? '';
-      const positionAccountId = indexedAccountId ?? otherWalletTypeAccountId;
-      const requestedScopes = [
-        {
-          stockOnly: options?.stockOnly,
-          supportNetworks,
-        },
-        ...(options?.additionalSupportNetworkScopes ?? []),
-      ]
-        .map((scope) => {
-          const scopeNetworks = Array.from(
-            new Map(
-              scope.supportNetworks
-                .filter((network) => Boolean(network.networkId))
-                .map((network) => [network.networkId, network]),
-            ).values(),
-          ).toSorted((left, right) =>
-            left.networkId.localeCompare(right.networkId),
-          );
-          const networkIdsKey = scopeNetworks
-            .map((network) => network.networkId)
-            .join(',');
-          return {
-            networkIds: new Set(
-              scopeNetworks.map((network) => network.networkId),
-            ),
-            networkIdsKey,
-            ownerKey: buildSwapProPositionsOwnerKey({
-              accountId: positionAccountId,
-              networkIdsKey,
-              currencyId: positionCurrencyId,
-              stockOnly: scope.stockOnly,
-            }),
-            stockOnly: Boolean(scope.stockOnly),
-            supportNetworks: scopeNetworks,
-          };
-        })
-        .filter((scope) => Boolean(scope.ownerKey));
-      const uniqueScopes = Array.from(
-        new Map(
-          requestedScopes.map((scope) => [scope.ownerKey, scope]),
-        ).values(),
-      );
-      if (uniqueScopes.length === 0) {
-        return;
-      }
-      const requestStartedAt = Date.now();
-      const runtimeEntries = get(swapProPositionsRuntimeDataAtom());
-      const activeRequestIds = get(swapProPositionsRequestIdsAtom());
-      const scopesToLoad = uniqueScopes.filter((scope) => {
-        if (!options?.forceRefresh && activeRequestIds[scope.ownerKey]) {
-          return false;
-        }
-        const runtimeEntry = runtimeEntries[scope.ownerKey];
-        return !(
-          !options?.forceRefresh &&
-          runtimeEntry?.status === 'success' &&
-          requestStartedAt - runtimeEntry.updatedAt <
-            SWAP_PRO_POSITIONS_RUNTIME_TTL_MS
-        );
-      });
-      if (scopesToLoad.length === 0) {
-        return;
-      }
-
-      const requestId = get(swapProPositionsRequestIdAtom()) + 1;
-      set(swapProPositionsRequestIdAtom(), requestId);
-      set(swapProPositionsRequestIdsAtom(), (previousRequestIds) => ({
-        ...previousRequestIds,
-        ...Object.fromEntries(
-          scopesToLoad.map((scope) => [scope.ownerKey, requestId]),
-        ),
-      }));
-      set(swapProPositionsRuntimeDataAtom(), (currentEntries) => {
-        let nextEntries = currentEntries;
-        for (const scope of scopesToLoad) {
-          nextEntries = upsertSwapProPositionsRuntimeEntry({
-            entries: nextEntries,
-            entry: {
-              status: 'loading',
-              tokens: [],
-              updatedAt: requestStartedAt,
-            },
-            ownerKey: scope.ownerKey,
-          });
-        }
-        return nextEntries;
-      });
-      const isLatestScopeRequest = (ownerKey: string) =>
-        get(swapProPositionsRequestIdsAtom())[ownerKey] === requestId;
-      const sortPositionTokens = (tokens: ISwapToken[]) =>
-        tokens.toSorted((left, right) =>
-          new BigNumber(right.fiatValue ?? '0').comparedTo(
-            new BigNumber(left.fiatValue ?? '0'),
-          ),
-        );
-      const publishNetworkPositions = ({
-        networkId,
-        scope,
-        tokens,
-      }: {
-        networkId: string;
-        scope: (typeof scopesToLoad)[number];
-        tokens: ISwapToken[];
-      }) => {
-        if (!isLatestScopeRequest(scope.ownerKey)) {
-          return;
-        }
-        set(swapProPositionsRuntimeDataAtom(), (currentEntries) => {
-          const currentEntry = currentEntries[scope.ownerKey];
-          if (!currentEntry) {
-            return currentEntries;
-          }
-          return upsertSwapProPositionsRuntimeEntry({
-            entries: currentEntries,
-            entry: {
-              status: 'loading',
-              tokens: sortPositionTokens([
-                ...currentEntry.tokens.filter(
-                  (token) => token.networkId !== networkId,
-                ),
-                ...tokens,
-              ]),
-              updatedAt: Date.now(),
-            },
-            ownerKey: scope.ownerKey,
-          });
-        });
-      };
-      try {
-        const unionSupportNetworks = Array.from(
-          new Map(
-            scopesToLoad
-              .flatMap((scope) => scope.supportNetworks)
-              .map((network) => [network.networkId, network]),
-          ).values(),
-        );
-        const {
-          supportAccountsFetchFailed,
-          swapSupportAccounts: swapProSupportAccounts,
-        } = await backgroundApiProxy.serviceSwap.getSupportSwapAllAccounts({
-          indexedAccountId,
-          otherWalletTypeAccountId,
-          swapSupportNetworks: unionSupportNetworks,
-        });
-        if (supportAccountsFetchFailed) {
-          set(swapProPositionsRuntimeDataAtom(), (currentEntries) => {
-            let nextEntries = currentEntries;
-            for (const scope of scopesToLoad) {
-              if (isLatestScopeRequest(scope.ownerKey)) {
-                const currentEntry = nextEntries[scope.ownerKey];
-                nextEntries = upsertSwapProPositionsRuntimeEntry({
-                  entries: nextEntries,
-                  entry: {
-                    status: 'error',
-                    tokens: currentEntry?.tokens ?? [],
-                    updatedAt: Date.now(),
-                  },
-                  ownerKey: scope.ownerKey,
-                });
-              }
-            }
-            return nextEntries;
-          });
-          return;
-        }
-        const accountAddressList = swapProSupportAccounts
-          .filter((item) => item.apiAddress)
-          .filter(
-            (item) => !networkUtils.isAllNetwork({ networkId: item.networkId }),
-          )
-          .toSorted((left, right) => {
-            const primaryNetworkIds = scopesToLoad[0]?.networkIds;
-            return (
-              Number(Boolean(primaryNetworkIds?.has(right.networkId))) -
-              Number(Boolean(primaryNetworkIds?.has(left.networkId)))
-            );
-          });
-        const stockNetworkIds = new Set(
-          scopesToLoad
-            .filter((scope) => scope.stockOnly)
-            .flatMap((scope) => [...scope.networkIds]),
-        );
-        const requestLocale = stockNetworkIds.size
-          ? appLocale.getLocale()
-          : undefined;
-        const tasks = accountAddressList.map((networkDataString) => {
-          const {
-            apiAddress,
-            networkId: accountNetworkId,
-            accountId,
-          } = networkDataString;
-          return async () => {
-            const tokens = await backgroundApiProxy.serviceSwap.fetchSwapTokens(
-              {
-                networkId: accountNetworkId,
-                accountNetworkId,
-                accountAddress: apiAddress,
-                accountId,
-                onlyAccountTokens: true,
-                isAllNetworkFetchAccountTokens: true,
-                throwOnError: true,
-                currency: positionCurrencyId,
-                protocol: ESwapTabSwitchType.SWAP,
-              },
-            );
-            for (const scope of scopesToLoad) {
-              if (!scope.stockOnly && scope.networkIds.has(accountNetworkId)) {
-                publishNetworkPositions({
-                  networkId: accountNetworkId,
-                  scope,
-                  tokens,
-                });
-              }
-            }
-
-            let stockMetadataFailed = false;
-            let stockTokens: ISwapToken[] | undefined;
-            if (stockNetworkIds.has(accountNetworkId)) {
-              try {
-                const stockCandidateTokens = tokens.filter((token) =>
-                  new BigNumber(token.fiatValue ?? '0').gt(
-                    swapProStockPositionsListMinValue,
-                  ),
-                );
-                if (stockCandidateTokens.length === 0) {
-                  stockTokens = [];
-                } else {
-                  const response =
-                    await backgroundApiProxy.serviceMarketV2.fetchMarketTokenListBatch(
-                      {
-                        requestLocale,
-                        tokenAddressList: stockCandidateTokens.map((token) => ({
-                          contractAddress: token.contractAddress ?? '',
-                          chainId: token.networkId,
-                          isNative: !!token.isNative,
-                        })),
-                      },
-                    );
-                  stockTokens = buildStockPositionTokens({
-                    marketItems: response.list ?? [],
-                    tokens: stockCandidateTokens,
-                  });
-                  if (!stockTokens) {
-                    throw new OneKeyLocalError(
-                      'Incomplete market metadata response for Stock positions',
-                    );
-                  }
-                }
-                for (const scope of scopesToLoad) {
-                  if (
-                    scope.stockOnly &&
-                    scope.networkIds.has(accountNetworkId)
-                  ) {
-                    publishNetworkPositions({
-                      networkId: accountNetworkId,
-                      scope,
-                      tokens: stockTokens,
-                    });
-                  }
-                }
-              } catch (error) {
-                stockMetadataFailed = true;
-                console.error('swapStock__loadPositionMetadata error', error);
-              }
-            }
-            return {
-              stockMetadataFailed,
-              stockTokens,
-              tokens,
-            };
-          };
-        });
-
-        // One union queue keeps iOS at three in-flight position requests even
-        // while the inactive Trade tabs are being warmed in the background.
-        const results = await this.executeBatched(tasks, 3);
-        for (const scope of scopesToLoad) {
-          if (isLatestScopeRequest(scope.ownerKey)) {
-            const scopeResults = results.filter((_, index) =>
-              scope.networkIds.has(accountAddressList[index]?.networkId ?? ''),
-            );
-            const hasFailure = scopeResults.some(
-              (result) =>
-                result.status === 'rejected' ||
-                (scope.stockOnly && result.value.stockMetadataFailed),
-            );
-            const scopeTokens = sortPositionTokens(
-              scopeResults.flatMap((result) => {
-                if (result.status !== 'fulfilled') {
-                  return [];
-                }
-                return scope.stockOnly
-                  ? (result.value.stockTokens ?? [])
-                  : result.value.tokens;
-              }),
-            );
-            const completedAt = Date.now();
-            set(swapProPositionsRuntimeDataAtom(), (currentEntries) =>
-              upsertSwapProPositionsRuntimeEntry({
-                entries: currentEntries,
-                entry: {
-                  status: hasFailure ? 'error' : 'success',
-                  tokens: scopeTokens,
-                  updatedAt: completedAt,
-                },
-                ownerKey: scope.ownerKey,
-              }),
-            );
-            if (!hasFailure) {
-              set(swapProPositionsCacheAtom(), (previousCache) =>
-                upsertSwapProPositionsCacheEntry({
-                  cache: getValidSwapProPositionsCache(previousCache),
-                  entry: {
-                    ownerKey: scope.ownerKey,
-                    networkIdsKey: scope.networkIdsKey,
-                    currencyId: positionCurrencyId,
-                    tokens: scopeTokens,
-                    updatedAt: completedAt,
-                  },
-                }),
-              );
-            }
-          }
-        }
-      } catch (error) {
-        set(swapProPositionsRuntimeDataAtom(), (currentEntries) => {
-          let nextEntries = currentEntries;
-          for (const scope of scopesToLoad) {
-            if (isLatestScopeRequest(scope.ownerKey)) {
-              const currentEntry = nextEntries[scope.ownerKey];
-              nextEntries = upsertSwapProPositionsRuntimeEntry({
-                entries: nextEntries,
-                entry: {
-                  status: 'error',
-                  tokens: currentEntry?.tokens ?? [],
-                  updatedAt: Date.now(),
-                },
-                ownerKey: scope.ownerKey,
-              });
-            }
-          }
-          return nextEntries;
-        });
-        console.error('swapPro__loadPositions error', error);
-      } finally {
-        set(swapProPositionsRequestIdsAtom(), (previousRequestIds) =>
-          Object.fromEntries(
-            Object.entries(previousRequestIds).filter(
-              ([ownerKey, activeRequestId]) =>
-                !scopesToLoad.some(
-                  (scope) =>
-                    scope.ownerKey === ownerKey &&
-                    activeRequestId === requestId,
-                ),
-            ),
-          ),
-        );
-      }
-    },
+    ) =>
+      options?.positionLoader?.(
+        get,
+        set,
+        this.executeBatched.bind(this),
+        supportNetworks,
+        indexedAccountId,
+        otherWalletTypeAccountId,
+        currencyId,
+        options,
+      ),
   );
 
   updateSwapProPositionTokenBalances = contextAtomMethod(
